functional_scripts/requirements.txt
aiohttp
asyncio
beautifulsoup4
fastapi
lxml
psycopg2-binary
requests
tqdm
tenacity
uvicorn